# api_paths/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnimalApiTestServer",
    "description": "Test Api Server for testing RestTOML",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18080",
      "description": "Animal Api Test Server"
    }
  ],
...
  "paths": {
    "/animal/delete/{id}": {
      "delete": {
        "summary": "Delete an animal",
        "operationId": "deleteAnimal",
        "tags": [
          "animals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Animal Id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An object that contains the message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpStatus"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/HttpStatus"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpStatus"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/HttpStatus"
                }
              }
            }
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpStatus"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/HttpStatus"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HttpStatus": {
        "type": "object",
        "properties": {
          "Status": {
            "type": "string"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "Message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}